backend/pyproject.toml
[project]
name = "helpey-backend"
version = "0.1.0"
description = "Helpey - Chat with your Google Drive"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "alembic>=1.14",
    # Config & Auth
    "pydantic-settings>=2.7",
    "pyjwt>=2.10",
    "cuid2>=2.0",
    # HTTP client
    "httpx>=0.28",
    # WorkOS SSO
    "workos>=5.0",
    # Task queue
    "celery>=5.4",
    # AI
    "anthropic>=0.52",
    "openai>=1.0",
    # Vector store
    "pgvector>=0.3",
    # Document processing
    "docling>=2.0",
    "chonkie[st]>=1.0",
    "pymupdf>=1.25",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "pillow>=11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
